split danger into many smaller files
one for each "task" from travis-land
