+++ danger/tasks/js-lint.js
@@ -1,0 +1,7 @@
+const {fileLog, readLogFile} = require('./logs.js')
+
+const eslintLog = readLogFile('./logs/eslint')
+
+if (eslintLog) {
+  fileLog('Eslint had a thing to say!', eslintLog)
+}
