+++ danger/tasks/js-prettier.js
@@ -1,0 +1,7 @@
+import {fileLog, readLogFile} from '../logs.js'
+
+const prettierLog = readLogFile('./logs/prettier')
+
+if (prettierLog) {
+  fileLog('Prettier made some changes', prettierLog, {lang: 'diff'})
+}
